refactored style folder
