+++ src/gui/pages/notifications_page.rs
@@ -36,8 +36,8 @@ pub fn notifications_page(sniffer: &Sniffer) -> Container<Message, Renderer<Styl
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
     let notifications = sniffer.settings.notifications;
-    let font = style.get_palette_extension().font;
-    let font_headers = style.get_palette_extension().font_headers;
+    let font = style.get_extension().font;
+    let font_headers = style.get_extension().font_headers;
 
     let mut tab_and_body = Column::new()
         .align_items(Alignment::Center)
@@ -398,7 +398,7 @@ fn get_button_clear_all(
 fn lazy_logged_notifications(sniffer: &Sniffer) -> Column<'static, Message, Renderer<StyleType>> {
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
-    let font = style.get_palette_extension().font;
+    let font = style.get_extension().font;
     let mut ret_val = Column::new()
         .width(Length::Fixed(830.0))
         .padding(5)
