+++ src/gui/pages/overview_page.rs
@@ -51,8 +51,8 @@ use crate::{AppProtocol, ChartType, Language, RunningPage, StyleType};
 pub fn overview_page(sniffer: &Sniffer) -> Container<Message, Renderer<StyleType>> {
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
-    let font = style.get_palette_extension().font;
-    let font_headers = style.get_palette_extension().font_headers;
+    let font = style.get_extension().font;
+    let font_headers = style.get_extension().font_headers;
 
     let mut body = Column::new();
     let mut tab_and_body = Column::new().height(Length::Fill);
@@ -242,7 +242,7 @@ fn lazy_row_report(sniffer: &Sniffer) -> Container<'static, Message, Renderer<St
 fn col_host(width: f32, sniffer: &Sniffer) -> Column<'static, Message, Renderer<StyleType>> {
     let language = sniffer.settings.language;
     let style = sniffer.settings.style;
-    let font = style.get_palette_extension().font;
+    let font = style.get_extension().font;
     let chart_type = sniffer.traffic_chart.chart_type;
 
     let mut scroll_host = Column::new()
@@ -339,7 +339,7 @@ fn col_host(width: f32, sniffer: &Sniffer) -> Column<'static, Message, Renderer<
 fn col_app(width: f32, sniffer: &Sniffer) -> Column<'static, Message, Renderer<StyleType>> {
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
-    let font = style.get_palette_extension().font;
+    let font = style.get_extension().font;
     let chart_type = sniffer.traffic_chart.chart_type;
 
     let mut col_app = Column::new()
@@ -414,7 +414,7 @@ fn lazy_col_info(
 ) -> Container<'static, Message, Renderer<StyleType>> {
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
-    let font = style.get_palette_extension().font;
+    let font = style.get_extension().font;
     let filtered_bytes =
         sniffer.runtime_data.tot_sent_bytes + sniffer.runtime_data.tot_received_bytes;
     let all_bytes = sniffer.runtime_data.all_bytes;
