+++ src/gui/pages/settings_general_page.rs
@@ -31,8 +31,8 @@ pub fn settings_general_page(sniffer: &Sniffer) -> Container<Message, Renderer<S
     let style = sniffer.settings.style;
     let language = sniffer.settings.language;
     let color_gradient = sniffer.settings.color_gradient;
-    let font = style.get_palette_extension().font;
-    let font_headers = style.get_palette_extension().font_headers;
+    let font = style.get_extension().font;
+    let font_headers = style.get_extension().font_headers;
 
     let content = Column::new()
         .align_items(Alignment::Center)
